variance bf16

+++ config/defaults.py
@@ -48,7 +48,7 @@ class train_config:
     # optimizer
     optimizer_type = "AnyPrecision"
     momentum_dtype = torch.float32
-    variance_dtype = torch.float32
+    variance_dtype = torch.bfloat16
     use_kahan = False
 
     # dataloaders
